Replace ShowAsPopupWithItems to CreatePopupWidgetWithInfo.  Also, make use of WebPopupMenuInfo from the WebKit API.  WebMenuItem remains in webkit/glue for convenience with IPC marshalling and related usage in Chrome.
This work is precursor to switching over to using WebWidget from
the WebKit API.

BUG=16234
TEST=html select drop downs should still work on the mac.  try
switching languages on news.google.com.
R=paul

Review URL: http://codereview.chromium.org/155378

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20624 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 88efb7ec99239eeecaa17d21f8635be1bce29cca

+++ webkit/glue/webmenurunner_mac.mm
@@ -25,7 +25,7 @@
     menu_.reset([[NSMenu alloc] initWithTitle:@""]);
     [menu_ setAutoenablesItems:NO];
     index_ = -1;
-    for (int i = 0; i < static_cast<int>(items.size()); ++i)
+    for (size_t i = 0; i < items.size(); ++i)
       [self addItem:items[i]];
   }
   return self;
